services/controller/pyproject.toml
[project]
name = "fixops-controller"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fixops-contract",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "langgraph>=0.2.55",
    "langgraph-checkpoint-postgres>=2.0.3",
    "langchain-core>=0.3.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29",
    "psycopg[binary]>=3.2",
    "pyyaml>=6.0",
    "pydantic-settings>=2.2",
]

[project.scripts]
fixops-controller = "fixops_controller.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
